+++ compiler/src/iree/compiler/Dialect/Stream/Transforms/BUILD.bazel
@@ -30,7 +30,6 @@ iree_compiler_cc_library(
         "ElideAsyncCopies.cpp",
         "ElideAsyncTransfers.cpp",
         "ElideTimepoints.cpp",
-        "ElideTransferOps.cpp",
         "EmplaceAllocations.cpp",
         "EncodeTensors.cpp",
         "FoldUniformOperands.cpp",
